Rename existing transactions UI components to Cosmos-
